Modularised code to increase cleanliness
