+++ src/pages/ArtifactView/index.js
@@ -1,9 +1,8 @@
 /* eslint-disable dot-notation */
 import React, { useEffect, useState } from 'react';
-import { Modal, Button } from 'react-bootstrap';
-import { useDropzone } from 'react-dropzone';
 import moment from 'moment';
 import PropTypes from 'prop-types';
+import PhotoForm from '../../components/photoform';
 import authFetchRequest from '../../utils/auth/cognitoFetchRequest';
 import styled from './index.module.scss';
 
@@ -14,9 +13,7 @@ function ArtifactView(props) {
   const [hasLoaded, setHasLoaded] = useState(false);
   const [errorState, setErrorState] = useState(false);
   const [showModal, setShowModal] = useState(false);
-  const { acceptedFiles, getRootProps, getInputProps } = useDropzone({
-    accept: 'image/jpeg, image/png'
-  });
+
   const { registerId, artifactId } = props.match.params;
 
   // TODO: write a hook to replicate useEffect authenticated fetch
@@ -50,39 +47,6 @@ function ArtifactView(props) {
   if (errorState) {
     return <div className="error">Something went wrong with your request, woops</div>;
   }
-
-  const addPhoto = () => {
-    setShowModal(false);
-    const promisedRequests = acceptedFiles.map(file => {
-      const formData = new FormData();
-      formData.append('registerId', registerId);
-      formData.append('artifactId', artifactId);
-      formData.append('photo', file);
-      return authFetchRequest('https://api.airloom.xyz/api/v1/artifact/addphoto/', {
-        method: 'POST',
-        body: formData
-      }).then(result => {
-        console.log(result);
-      });
-    });
-
-    Promise.all(promisedRequests)
-      .then(results => {
-        console.log(results);
-      })
-      .catch(err => {
-        console.log(err);
-      });
-  };
-
-  const files = acceptedFiles.map(file => (
-    <li key={file.path}>
-      {file.path} -{file.size} bytes
-    </li>
-  ));
-
-  const submitButtonClass =
-    acceptedFiles.length > 0 ? styled['submit-button--enabled'] : styled['submit-button--disabled'];
 
   return (
     <>
@@ -138,35 +102,12 @@ function ArtifactView(props) {
           </div>
         </div>
       </div>
-      <Modal show={showModal} onHide={() => setShowModal(false)}>
-        <Modal.Header closeButton>Add a photo</Modal.Header>
-        <Modal.Body>
-          <section className="container">
-            <div {...getRootProps({ className: styled['dropzone'] })}>
-              <input {...getInputProps()} />
-              <p>Drag 'n' drop some files here, or click to select files</p>
-              <em>(Only *.jpeg and *.png images will be accepted)</em>
-            </div>
-            <aside>
-              <h4>Files</h4>
-              <ul>{files}</ul>
-            </aside>
-          </section>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button
-            variant="secondary"
-            onClick={() => {
-              setShowModal(false);
-            }}
-          >
-            Close
-          </Button>
-          <Button className={submitButtonClass} onClick={addPhoto} variant="primary">
-            Add a photo
-          </Button>
-        </Modal.Footer>
-      </Modal>
+      <PhotoForm
+        registerId={registerId}
+        artifactId={artifactId}
+        showModal={showModal}
+        setShowModal={setShowModal}
+      />
     </>
   );
 }
